Perpl_LongTerm_Server/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.28
protobuf==3.19.3
requests==2.27.1
six==1.16.0
spotipy==2.19.0
urllib3==1.26.8
Werkzeug==2.0.2
youtube-dl==2021.12.17
